add cuda test

+++ discretelog.cu
@@ -42,7 +42,7 @@ __global__ void sumCommMultiBlock(
     const int gridSize = blockSize * gridDim.x;
 
     int sum = 0;
-    int res = 0;
+    long long int res = 0;
     for (int i = start + gthIdx; i < end; i += gridSize)
         {
             res = pow(base, i);
